refactor: изменить тип поля dto
